move win tools updater to storage
